+++ src/data/en/index.ts
@@ -9,6 +9,7 @@
 import type { Bundle } from '../zh';
 
 import actions from './actions.json';
+import codex from './codex.json';
 import opening from './opening.json';
 import systemLines from './system_lines.json';
 import ui from './ui.json';
@@ -65,6 +66,7 @@ import eventsRandomRandomWell from './events/random/random_well.json';
 
 const bundle: Bundle = {
   actions: actions,
+  codex: codex,
   opening: opening,
   systemLines: systemLines,
   ui: ui,